refactor(tests): split test_doctor.py into three focused files (P1-F02)
Extracts TestDoctorScriptHealth + TestSyncRemovalCLI into test_doctor_scripts.py
and the full migration/fleet/feature-gate section (lines 1503–2530) into
test_doctor_migration.py. Adds test_doctor_split.py structural guards. All three
files carry the required pytestmark = [pytest.mark.layer("cli"), pytest.mark.small].
