+++ tests/cli/test_doctor_scripts.py
@@ -1,0 +1,257 @@
+"""Tests for doctor script/recipe version health checks."""
+
+from __future__ import annotations
+
+import json
+from pathlib import Path
+
+import pytest
+
+from autoskillit import cli
+
+pytestmark = [pytest.mark.layer("cli"), pytest.mark.small]
+
+_MINIMAL_SCRIPT_YAML = """\
+name: my-script
+description: A test script
+steps:
+  do_it:
+    tool: run_cmd
+    with:
+      cmd: echo hello
+    on_success: done
+  done:
+    action: stop
+    message: Done
+kitchen_rules:
+  - Only use AutoSkillit MCP tools during pipeline execution
+"""
+
+
+class TestDoctorScriptHealth:
+    """Doctor check for script version staleness."""
+
+    # DOC1: No .autoskillit/recipes/ -> OK result
+    def test_no_scripts_dir_reports_ok(
+        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture
+    ) -> None:
+        """doctor reports OK for script_version_health when no scripts directory exists."""
+        monkeypatch.setattr(Path, "home", lambda: tmp_path)
+        monkeypatch.chdir(tmp_path)
+        # No .autoskillit/recipes/ directory created
+        cli.doctor(output_json=True)
+        captured = capsys.readouterr()
+        data = json.loads(captured.out)
+        script_checks = [r for r in data["results"] if r["check"] == "script_version_health"]
+        assert len(script_checks) == 1
+        assert script_checks[0]["severity"] == "ok"
+
+    # DOC2: All scripts at current version -> OK result
+    def test_all_scripts_at_current_version_reports_ok(
+        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture
+    ) -> None:
+        """doctor reports OK when all scripts carry the current installed version."""
+        import autoskillit
+
+        current_version = autoskillit.__version__
+        monkeypatch.setattr(Path, "home", lambda: tmp_path)
+        monkeypatch.chdir(tmp_path)
+        scripts_dir = tmp_path / ".autoskillit" / "recipes"
+        scripts_dir.mkdir(parents=True)
+        (scripts_dir / "up-to-date.yaml").write_text(
+            f"name: up-to-date\ndescription: Current version\n"
+            f'autoskillit_version: "{current_version}"\n'
+            + _MINIMAL_SCRIPT_YAML.split("\n", 2)[2]  # reuse steps/constraints block
+        )
+        cli.doctor(output_json=True)
+        captured = capsys.readouterr()
+        data = json.loads(captured.out)
+        script_checks = [r for r in data["results"] if r["check"] == "script_version_health"]
+        assert len(script_checks) == 1
+        assert script_checks[0]["severity"] == "ok"
+
+    # DOC3: Scripts below current version -> WARNING result with recipe names
+    def test_outdated_scripts_reports_warning_with_count(
+        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture
+    ) -> None:
+        """doctor reports WARNING with recipe names when scripts have an older version."""
+        import autoskillit
+
+        monkeypatch.setattr(autoskillit, "__version__", "99.0.0")
+        monkeypatch.setattr(Path, "home", lambda: tmp_path)
+        monkeypatch.chdir(tmp_path)
+        scripts_dir = tmp_path / ".autoskillit" / "recipes"
+        scripts_dir.mkdir(parents=True)
+        (scripts_dir / "old-script.yaml").write_text(
+            'name: old-script\ndescription: Old\nautoskillit_version: "0.1.0"\n'
+        )
+        (scripts_dir / "also-old.yaml").write_text(
+            'name: also-old\ndescription: Also old\nautoskillit_version: "0.1.0"\n'
+        )
+        cli.doctor(output_json=True)
+        captured = capsys.readouterr()
+        data = json.loads(captured.out)
+        script_checks = [r for r in data["results"] if r["check"] == "script_version_health"]
+        assert len(script_checks) == 1
+        assert script_checks[0]["severity"] == "warning"
+        assert "old-script" in script_checks[0]["message"]
+
+    # DOC4: Scripts with no version field -> WARNING result
+    def test_scripts_without_version_reports_warning(
+        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture
+    ) -> None:
+        """doctor reports WARNING when script YAML has no autoskillit_version field."""
+        monkeypatch.setattr(Path, "home", lambda: tmp_path)
+        monkeypatch.chdir(tmp_path)
+        scripts_dir = tmp_path / ".autoskillit" / "recipes"
+        scripts_dir.mkdir(parents=True)
+        (scripts_dir / "no-version.yaml").write_text(
+            "name: no-version\ndescription: No version field\n"
+        )
+        cli.doctor(output_json=True)
+        captured = capsys.readouterr()
+        data = json.loads(captured.out)
+        script_checks = [r for r in data["results"] if r["check"] == "script_version_health"]
+        assert len(script_checks) == 1
+        assert script_checks[0]["severity"] == "warning"
+
+    def _setup_recipe(self, scripts_dir: Path, name: str, version: str = "0.1.0") -> None:
+        (scripts_dir / f"{name}.yaml").write_text(
+            f'name: {name}\ndescription: Test\nautoskillit_version: "{version}"\n'
+        )
+
+    def _write_failures_json(self, tmp_path: Path, name: str, retries: int = 3) -> None:
+        import json as _json
+
+        failures_path = tmp_path / ".autoskillit" / "temp" / "migrations" / "failures.json"
+        failures_path.parent.mkdir(parents=True, exist_ok=True)
+        failures_path.write_text(
+            _json.dumps(
+                {
+                    name: {
+                        "name": name,
+                        "file_path": f"/fake/{name}.yaml",
+                        "file_type": "recipe",
+                        "timestamp": "2026-01-01T00:00:00+00:00",
+                        "error": "validation failed after retries",
+                        "retries_attempted": retries,
+                    }
+                }
+            )
+        )
+
+    # DR1: failures.json has an entry for a recipe -> error severity
+    def test_doctor_error_on_failed_migration(
+        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture
+    ) -> None:
+        """doctor reports error severity when failures.json has an entry for a recipe."""
+        import autoskillit
+
+        monkeypatch.setattr(autoskillit, "__version__", "99.0.0")
+        monkeypatch.setattr(Path, "home", lambda: tmp_path)
+        monkeypatch.chdir(tmp_path)
+        scripts_dir = tmp_path / ".autoskillit" / "recipes"
+        scripts_dir.mkdir(parents=True)
+        self._setup_recipe(scripts_dir, "broken-pipeline")
+        self._write_failures_json(tmp_path, "broken-pipeline", retries=3)
+        cli.doctor(output_json=True)
+        captured = capsys.readouterr()
+        data = json.loads(captured.out)
+        script_checks = [r for r in data["results"] if r["check"] == "script_version_health"]
+        assert len(script_checks) == 1
+        assert script_checks[0]["severity"] == "error"
+
+    # DR2: Error message includes retries_attempted value from failure record
+    def test_doctor_error_message_includes_retry_count(
+        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture
+    ) -> None:
+        """doctor error message includes retries_attempted value from the failure record."""
+        import autoskillit
+
+        monkeypatch.setattr(autoskillit, "__version__", "99.0.0")
+        monkeypatch.setattr(Path, "home", lambda: tmp_path)
+        monkeypatch.chdir(tmp_path)
+        scripts_dir = tmp_path / ".autoskillit" / "recipes"
+        scripts_dir.mkdir(parents=True)
+        self._setup_recipe(scripts_dir, "my-pipeline")
+        self._write_failures_json(tmp_path, "my-pipeline", retries=3)
+        cli.doctor(output_json=True)
+        captured = capsys.readouterr()
+        data = json.loads(captured.out)
+        script_checks = [r for r in data["results"] if r["check"] == "script_version_health"]
+        assert len(script_checks) == 1
+        assert "3" in script_checks[0]["message"]
+
+    # DR3: Outdated recipe with no failure record -> warning severity
+    def test_doctor_warning_on_simply_outdated(
+        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture
+    ) -> None:
+        """doctor reports warning when recipe is outdated but has no failure record."""
+        import autoskillit
+
+        monkeypatch.setattr(autoskillit, "__version__", "99.0.0")
+        monkeypatch.setattr(Path, "home", lambda: tmp_path)
+        monkeypatch.chdir(tmp_path)
+        scripts_dir = tmp_path / ".autoskillit" / "recipes"
+        scripts_dir.mkdir(parents=True)
+        self._setup_recipe(scripts_dir, "outdated-pipeline")
+        # No failures.json written
+        cli.doctor(output_json=True)
+        captured = capsys.readouterr()
+        data = json.loads(captured.out)
+        script_checks = [r for r in data["results"] if r["check"] == "script_version_health"]
+        assert len(script_checks) == 1
+        assert script_checks[0]["severity"] == "warning"
+
+    # DR4: All recipes current, no failures.json -> ok
+    def test_doctor_ok_when_all_current(
+        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture
+    ) -> None:
+        """doctor reports ok when all recipes are at current version and no failures.json."""
+        import autoskillit
+
+        current_version = autoskillit.__version__
+        monkeypatch.setattr(Path, "home", lambda: tmp_path)
+        monkeypatch.chdir(tmp_path)
+        scripts_dir = tmp_path / ".autoskillit" / "recipes"
+        scripts_dir.mkdir(parents=True)
+        self._setup_recipe(scripts_dir, "current-pipeline", version=current_version)
+        # No failures.json written
+        cli.doctor(output_json=True)
+        captured = capsys.readouterr()
+        data = json.loads(captured.out)
+        script_checks = [r for r in data["results"] if r["check"] == "script_version_health"]
+        assert len(script_checks) == 1
+        assert script_checks[0]["severity"] == "ok"
+
+    # DR5: Warning message says "Will be auto-migrated on next load"
+    def test_doctor_outdated_message_updated(
+        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture
+    ) -> None:
+        """doctor warning message says 'Will be auto-migrated on next load'."""
+        import autoskillit
+
+        monkeypatch.setattr(autoskillit, "__version__", "99.0.0")
+        monkeypatch.setattr(Path, "home", lambda: tmp_path)
+        monkeypatch.chdir(tmp_path)
+        scripts_dir = tmp_path / ".autoskillit" / "recipes"
+        scripts_dir.mkdir(parents=True)
+        self._setup_recipe(scripts_dir, "stale-pipeline")
+        cli.doctor(output_json=True)
+        captured = capsys.readouterr()
+        data = json.loads(captured.out)
+        script_checks = [r for r in data["results"] if r["check"] == "script_version_health"]
+        assert len(script_checks) == 1
+        assert script_checks[0]["severity"] == "warning"
+        assert "Will be auto-migrated on next load" in script_checks[0]["message"]
+
+
+class TestSyncRemovalCLI:
+    def test_doctor_has_no_recipe_sync_check(
+        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture
+    ):
+        """REQ-APP-006: doctor output does not include recipe_sync_status."""
+        monkeypatch.chdir(tmp_path)
+        cli.doctor()
+        captured = capsys.readouterr()
+        assert "recipe_sync_status" not in captured.out
